requirements.txt
autopep8==1.6.0
awscrt==0.13.14
boto3==1.24
boto3-stubs==1.24
botocore==1.27
botocore-stubs==1.27
certifi==2022.6.15
chardet==5.0.0
jmespath==1.0.1
lxml==4.9.1
mypy-boto3-s3==1.24
Pillow==9.2.0
pycodestyle==2.8.0
PyPDF2==2.8.1
python-dateutil==2.8.2
python-magic==0.4.27
s3transfer==0.6.0
six==1.16.0
toml==0.10.2
types-chardet==5.0.3
types-lxml==2022.4.10
types-Pillow==9.2.0
typing_extensions==4.3.0
urllib3==1.26.11
